[wip] Works around ROOT limitations with separate single and multi layout collections
ROOT does not manage the case where the multi collection only has one layout.
This is worked around by keeping the previous code for the single layout collection
and creating the multi layout collections for the case where there are at least two layouts.
It would work with only one, but ROOT fails to compile it.